agents/Clipy/requirements.txt
cookiecutter==1.7.0
unstructured
python-dotenv
llama-index-readers-github
llama-index-embeddings-huggingface
llama-index-readers-youtube-transcript

# comment out the line below if you do not have a GPU
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio 